refactor(web): consolidate API utilities into data layer
Move utils/api/cars.ts and utils/api/coe.ts into lib/data/ following Next.js patterns. Separate pure utility functions into lib/utils/ to comply with server action requirements.
