+++ apps/web/src/app/(dashboard)/cars/[category]/category-tabs.tsx
@@ -8,11 +8,11 @@ import {
   MarketShareDonut,
   TopPerformersBar,
 } from "@web/components/charts";
-import type { TypeItem } from "@web/types";
 import type {
   CarMarketShareResponse,
   CarTopPerformersData,
-} from "@web/utils/api/cars";
+} from "@web/lib/data/cars";
+import type { TypeItem } from "@web/types";
 import { formatPercentage } from "@web/utils/chart-formatters";
 import { formatDateToMonthYear } from "@web/utils/format-date-to-month-year";
 import { Suspense } from "react";
