+++ apps/web/src/lib/data/cars.ts
@@ -1,11 +1,161 @@
 "use server";
 
+import { API_URL } from "@web/config";
+import { CAR_CHART_COLOURS } from "@web/lib/utils/cars";
+import { RevalidateTags } from "@web/types";
 import type {
-  CarComparisonMetrics,
-  CarMarketShareResponse,
-  CarTopPerformersData,
-} from "@web/utils/api/cars";
-import { getCarsData, getTopMakes, getTopTypes } from "@web/utils/api/cars";
+  Comparison,
+  FuelType,
+  Registration,
+  TopType,
+} from "@web/types/cars";
+import { CACHE_DURATION, fetchApi } from "@web/utils/fetch-api";
+
+// Cars-specific types
+export interface CarMarketShareData {
+  name: string;
+  count: number;
+  percentage: number;
+  colour: string;
+}
+
+export interface CarMarketShareResponse {
+  month: string;
+  total: number;
+  category: "fuelType" | "vehicleType";
+  data: CarMarketShareData[];
+  dominantType: {
+    name: string;
+    percentage: number;
+  };
+}
+
+export interface CarTopTypeData {
+  name: string;
+  count: number;
+  percentage: number;
+  rank: number;
+}
+
+export interface CarTopMakeData {
+  make: string;
+  count: number;
+  percentage: number;
+  rank: number;
+  fuelType?: string;
+  vehicleType?: string;
+}
+
+export interface CarTopPerformersData {
+  month: string;
+  total: number;
+  topFuelTypes: CarTopTypeData[];
+  topVehicleTypes: CarTopTypeData[];
+  topMakes: CarTopMakeData[];
+}
+
+export interface CarComparisonData {
+  month: string;
+  total: number;
+  fuelType: Array<{ name: string; count: number }>;
+  vehicleType: Array<{ name: string; count: number }>;
+}
+
+export interface CarComparisonMetrics {
+  currentMonth: CarComparisonData;
+  previousMonth: CarComparisonData;
+  growthRate: number;
+  monthOverMonth: {
+    total: number;
+    fuelTypes: Array<{
+      name: string;
+      current: number;
+      previous: number;
+      growth: number;
+    }>;
+    vehicleTypes: Array<{
+      name: string;
+      current: number;
+      previous: number;
+      growth: number;
+    }>;
+  };
+}
+
+export const getCarsData = async (month: string): Promise<Registration> => {
+  return fetchApi<Registration>(`${API_URL}/cars?month=${month}`, {
+    next: {
+      tags: [RevalidateTags.Cars, `${RevalidateTags.Cars}:${month}`],
+      revalidate: CACHE_DURATION,
+    },
+  });
+};
+
+export const getCarsComparison = async (month: string): Promise<Comparison> => {
+  return fetchApi<Comparison>(`${API_URL}/cars/compare?month=${month}`, {
+    next: {
+      tags: [
+        RevalidateTags.Analysis,
+        RevalidateTags.Comparison,
+        RevalidateTags.Cars,
+      ],
+      revalidate: CACHE_DURATION,
+    },
+  });
+};
+
+export const getTopTypes = async (month: string): Promise<TopType> => {
+  return fetchApi<TopType>(`${API_URL}/cars/top-types?month=${month}`, {
+    next: {
+      tags: [
+        RevalidateTags.Analysis,
+        RevalidateTags.TopPerformers,
+        RevalidateTags.Cars,
+      ],
+      revalidate: CACHE_DURATION,
+    },
+  });
+};
+
+export const getTopMakes = async (month: string): Promise<FuelType[]> => {
+  return fetchApi<FuelType[]>(`${API_URL}/cars/top-makes?month=${month}`, {
+    next: {
+      tags: [
+        RevalidateTags.Analysis,
+        RevalidateTags.TopPerformers,
+        RevalidateTags.Cars,
+      ],
+      revalidate: CACHE_DURATION,
+    },
+  });
+};
+
+export const getMakesList = async (): Promise<string[]> => {
+  return fetchApi<string[]>(`${API_URL}/cars/makes`, {
+    next: {
+      tags: [RevalidateTags.Reference, RevalidateTags.Cars],
+      revalidate: CACHE_DURATION,
+    },
+  });
+};
+
+export const getLatestCarMonth = async (): Promise<{ month: string }> => {
+  return fetchApi<{ month: string }>(`${API_URL}/cars/months/latest`, {
+    next: {
+      tags: [RevalidateTags.Latest, RevalidateTags.Reference, "cars"],
+      revalidate: CACHE_DURATION,
+    },
+  });
+};
+
+export const getCarMonthsList = async (): Promise<string[]> => {
+  return fetchApi<string[]>(`${API_URL}/cars/months`, {
+    next: {
+      tags: [RevalidateTags.Reference, "cars"],
+      revalidate: CACHE_DURATION,
+    },
+  });
+};
 
 export const getCarMarketShareData = async (
   month: string,
