+++ apps/web/src/utils/cached-api.ts
@@ -2,48 +2,23 @@ export {
   getCarComparisonData,
   getCarHistoricalTrends,
   getCarMarketShareData,
+  getCarMonthsList,
+  getCarsComparison,
+  getCarsData,
   getCarTopPerformersData,
+  getLatestCarMonth,
+  getMakesList,
+  getTopMakes,
+  getTopTypes,
 } from "@web/lib/data/cars";
 export {
   getCOEComparisonData,
   getCOEMarketShareData,
-  getCOETrendData,
-} from "@web/lib/data/coe";
-export {
-  CAR_CHART_COLOURS,
-  type CarComparisonData,
-  type CarComparisonMetrics,
-  type CarMarketShareData,
-  type CarMarketShareResponse,
-  type CarTopMakeData,
-  type CarTopPerformersData,
-  type CarTopTypeData,
-  calculateCarMarketShareInsights,
-  formatCarMarketShareForChart,
-  getCarMonthsList,
-  getCarsComparison,
-  getCarsData,
-  getLatestCarMonth,
-  getMakesList,
-  getRankingEmoji,
-  getTopMakes,
-  getTopTypes,
-} from "./api/cars";
-export {
-  COE_CATEGORY_MAP,
-  COE_CHART_COLOURS,
-  type COEComparisonMetrics,
-  type COEMarketShareData,
-  type COEMarketShareResponse,
-  type COETrendData,
-  calculateCOEPremiumInsights,
-  formatCOECategory,
-  formatCOEDataForChart,
-  getCOECategoryColor,
   getCOEMonthsList,
   getCOEResults,
   getCOEResultsByPeriod,
+  getCOETrendData,
   getLatestCOEMonth,
   getLatestCOEResults,
   getPQPData,
-} from "./api/coe";
+} from "@web/lib/data/coe";
